requirements.txt
streamlit
pandas
numpy
plotly
altair
matplotlib
openai
google-generativeai 
python-dotenv